+++ sql/schema/stats/moveset_rated_teammates.sql
@@ -1,7 +1,7 @@
 create table if not exists `moveset_rated_teammates`
 (
 `usage_rated_pokemon_id` int unsigned not null,
-`teammate_id` smallint unsigned not null,
+`teammate_id` mediumint unsigned not null,
 
 `percent` decimal(6, 3) not null,
 # `percent` % of teams that have `pokemon_id` also have `teammate_id`.
